MNT Move min_dependencies up a level (#18610)
