+++ build_tools/shared.sh
@@ -11,6 +11,6 @@ get_dep() {
         # use latest
         echo "$package"
     elif [[ "$version" == "min" ]]; then
-        echo "$package==$(python sklearn/_build_utils/min_dependencies.py $package)"
+        echo "$package==$(python sklearn/_min_dependencies.py $package)"
     fi
 }
