+++ setup.py
@@ -50,7 +50,7 @@ PROJECT_URLS = {
 # We can actually import a restricted version of sklearn that
 # does not need the compiled code
 import sklearn
-import sklearn._build_utils.min_dependencies as min_deps  # noqa
+import sklearn._min_dependencies as min_deps  # noqa
 
 
 VERSION = sklearn.__version__
